web_py/requirements.txt
beautifulsoup4==4.9.3
click==7.1.2
feedparser==6.0.2
Flask==1.1.2
Flask-Cors==3.0.10
google-cloud==0.34.0
gunicorn==20.1.0
html5lib==1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.2.1
tabulate==0.8.9
webencodings==0.5.1
Werkzeug==1.0.1
